Portal Spawn Checker
Added a class to check the potential spawn position of portals. Currently only ensures the portal is not buried inside a wall. Next step is to prevent portals from overlapping each other, and to prevent a portal spawning if it is not fully covered on the back by a surface.